Refactor CMake code related to compiler warnings, enable some additional warnings (#6355)

+++ src/extractor/extraction_containers.cpp
@@ -667,7 +667,7 @@ void ExtractionContainers::PrepareEdges(ScriptingEnvironment &scripting_environm
             auto &edge = edge_iterator->result;
             edge.weight = std::max<EdgeWeight>(1, std::round(segment.weight * weight_multiplier));
             edge.duration = std::max<EdgeWeight>(1, std::round(segment.duration * 10.));
-            edge.distance = accurate_distance;
+            edge.distance = static_cast<float>(accurate_distance);
 
             // assign new node id
             const auto node_id = mapExternalToInternalNodeID(
